Revert "feat(St-22): implement authguard (#3)"
This reverts commit c64f4f44b08e06516610f206891e34c80b3547d7.
