+++ package.json
@@ -54,7 +54,6 @@
     "react-hook-form-mantine": "^2.0.0",
     "react-icons": "^4.12.0",
     "react-redux": "^8.0.5",
-    "redux-persist": "^6.0.0",
     "socket.io-client": "^4.7.5",
     "uuid": "^9.0.1",
     "zod": "^3.22.4"
