+++ shared/constants/route.constants.ts
@@ -1,2 +1,2 @@
-export const HOME_ROUTE = "/dashboard/student";
-export const SUPER_USER_DASHBOARD_ROUTE = "/dashboard/teacher";
+export const HOME_ROUTE = "/";
+export const SUPER_USER_DASHBOARD_ROUTE = "/super-user-dashboard";
