--- shared/layout/dashboard.layout.tsx
@@ -1,10 +1,0 @@
-import Navbar from "@/shared/components/Navbar";
-
-export default function DashboardLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <div>
-      <Navbar />
-      {children}
-    </div>
-  );
-}
