+++ shared/redux/rtk-apis/auth/auth.types.ts
@@ -6,7 +6,6 @@ export type TLoginRequestFields = {
 export enum EUserRole {
   STUDENT = "STUDENT",
   TEACHER = "TEACHER",
-  NONE = "NONE",
 }
 
 export type TTokenizedUser = {
