+++ shared/redux/store.ts
@@ -1,31 +1,17 @@
-import { configureStore, combineReducers } from "@reduxjs/toolkit";
+import { configureStore } from "@reduxjs/toolkit";
 import { setupListeners } from "@reduxjs/toolkit/query/react";
-import { persistStore, persistReducer } from "redux-persist";
-import storage from "redux-persist/lib/storage";
 
 import authenticatedUserSliceReducer from "./reducers/user.reducer";
 import projectApi from "./rtk-apis/api.config";
 
-const rootReducer = combineReducers({
-  authenticatedUser: authenticatedUserSliceReducer,
-  [projectApi.reducerPath]: projectApi.reducer,
-});
-
-const persistConfig = {
-  key: "root",
-  storage,
-  whitelist: ["authenticatedUser"],
-};
-
-const persistedReducer = persistReducer(persistConfig, rootReducer);
-
 export const store = configureStore({
-  reducer: persistedReducer,
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware({ serializableCheck: false }).concat(projectApi.middleware),
-});
+  reducer: {
+    authenticatedUser: authenticatedUserSliceReducer,
+    [projectApi.reducerPath]: projectApi.reducer,
+  },
 
-export const persistor = persistStore(store);
+  middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(projectApi.middleware),
+});
 
 setupListeners(store.dispatch);
